Add surjection mapping container
Add a surjective mapping container, composed of a sequence of pairs.
This type provides lookup of a value type given a key type. A default
type is returned on lookup failure.

This commit adds unit tests but also "compilation failure" tests. This
ensures that the surjective class invariants, specified with
`static_assert`, cause compilation failures if violated.