Add Variant class and namespace
Add Variant class to represent a type-safe union. This class is defined
in the `variant` namespace, along with helper classes
`bad_variant_access` and `empty`.

The behavior of Variant is similar to std::variant (since C++17), but
always contains `empty` as the first type and allows the alternative
type held within the object to be taken out. When the alternative type
is taken, it is replaced with `empty`.