Add Transition class
Add Transition class to represent a state machine transition for a given
state and event pair. A transition can contain a guard, which determines
if a transition is triggered. On trigger, an associated action is
invoked which creates the destination state.

When creating transitions, an empty placeholder can be used as a
substitute for default values. An empty guard is converted into a guard
function that always triggers the transition. An empty destination
represents an internal state transition[1]. An empty action can be used
with an empty destination.

[1]: https://en.wikipedia.org/wiki/UML_state_machine#Internal_transitions