Don't inherit from std::tuple in transition Row
If transition::Row inherits from std::tuple and (in a later commit)
transition::Table inherits from std::tuple, std::get may fail to type
deduction due to ambiguity[1].

This commit updates transition Row to contain a std::tuple as a private
data member instead of inheriting from std::tuple. As accessor is
provided for lookup. An rvalue reference qualified member function is
provided for extracting the tuple in order to update a Row with a new
Transition.

[1]: https://github.com/microsoft/STL/issues/121