Add transition Table class
Add a transition Table class which is composed of a collection of
transition Rows. A Table can be created from multiple Rows or updated
with a new Transition.

On update, the new Transition is placed in a new Row within the Table if
the key is new. If the key is already associated with a specific Row,
the Transition is appended to that Row.