Add StateMachine class
Define a StateMachine class that is composed of a transition Table. The
initial state is determined by the first item in the state list.

The destination_types are defined for a transition Row and the
event_types and state_types are defined for a transition Table, in order
to allow a StateMachine to determine valid states and events.