Add index method to Variant
Add a method to obtain the current index of the held type in a Variant,
allowing a StateMachine to inspect the type. This commit also exposes
the type aliases for `index_type`, `alternative_index_map`, method
`alternative_index()`, and defines a `size` member type for the number
of alternative types.