Define an optional type
A number of find operations return an index, but can fail. On failure, a
'sentinel/null' value is used. In order to improve type semantics, these
return types are replaced with an optional type.

Since C++14 doesn't offer an optional type, a simple one is provided
that implements a subset of std::optional.