Perform state transition in process event
On StateMachine::process_event, find corresponding transition row. For
transitions in a row, invoke the transition guard and perform the
associated action if the guard succeeds (return true). If the transition
is internal, on exit and on entry actions will be skipped. If not, the
transition will consist of:
- on exit action of the source state,
- transition action creating the destination state
- on entry action of the destination state
and with an update of the state.

Transition guards are evaluated in storage order and there is no
mechanism to determine if multiple transition guards will succeed.
Processing stops as soon as the first guard succeeds and the action is
performed.