Define Bazel configuration for examples and tests
Define this project as a Bazel workspace and build files for the state
machine library, examples, and tests.

A separate `generated_compiler_config` workspace is defined in order to
detect the CXX compiler without requiring explicit user specification.
This generates the default COPTS for the library and uses different
compiler warning options depending on use of Clang or GCC.

This commit only defines unit tests using Googletest. It does not port
the "try-compile" tests from CMake.