Define a lint macro and lint examples
This commit defines a macro for linting source files where each file is
linted as a separate action. A default clang-tidy config is defined for
the repository and used by this macro.

The clang-tidy `InheritParentConfig` feature is only available in
clang-tidy 12, however the `-checks` option can be used to append checks
to the default config.

This commit also enables linting for examples.