Add `lint` and `manual` tags to lint rules
Normally when building in CI, we don't want to lint all source files.
Linting requires the same binary and would result in repeated work with
multiple compiler configurations. As a result, the lint rules are tagged
as `manual` to prevent inclusion in target pattern wildcards.

The `lint` tag is also added to allow filtering with `bazel query`. This
allows all lint test targets to be determined with the following:

    bazel query 'attr(tags, "lint", tests(//...))'